+++ orders/urls.py
@@ -1,7 +1,7 @@
-from django.urls import path
-
-from orders.views import download_excel_file
-
-urlpatterns = [
-    path('download/', download_excel_file, name='download_excel')
-]
+# from django.urls import path
+#
+# from orders.views import download_excel_file
+#
+# urlpatterns = [
+#     path('download/', download_excel_file, name='download_excel')
+# ]
